requirements.txt
# NextCare Application Dependencies
# Core web framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1

# Database
psycopg2-binary==2.9.7
alembic==1.12.0

# Authentication and security
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0

# Data processing and validation
pandas==2.1.1
numpy==1.24.3
marshmallow==3.20.1

# API and serialization
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
jsonschema==4.19.1

# Background tasks and scheduling
celery==5.3.2
redis==4.6.0
APScheduler==3.10.4

# Monitoring and logging
prometheus-client==0.17.1
structlog==23.1.0

# Email notifications
Flask-Mail==0.9.1

# Mock sensor simulation
requests==2.31.0
Faker==19.6.2

# Development and testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
factory-boy==3.3.0
freezegun==1.2.2

# Code quality
black==23.7.0
flake8==6.0.0
pylint==2.17.5
mypy==1.5.1

# Documentation
Sphinx==7.1.2
sphinx-rtd-theme==1.3.0

# Deployment
gunicorn==21.2.0
supervisor==4.2.5

# Environment and configuration
click==8.1.7
python-decouple==3.8

# Date and time utilities
python-dateutil==2.8.2
pytz==2023.3

# File handling
openpyxl==3.1.2
Pillow==10.0.0

# WebSocket support for real-time updates
Flask-SocketIO==5.3.6
python-socketio==5.8.0

# Caching
Flask-Caching==2.1.0

# Rate limiting
Flask-Limiter==3.5.0

# Health checks
Flask-HealthCheck==1.0.0

# Metrics and monitoring
statsd==4.0.1